+++ fsvolumelist/page_filterdriver.h
@@ -149,7 +149,7 @@ public:
 
 			hwndList = CreateWindowEx(0,WC_LISTVIEW, 
                               L"",
-                              WS_CHILD | WS_CLIPCHILDREN | LVS_REPORT | LVS_SHOWSELALWAYS,
+                              WS_CHILD | WS_CLIPCHILDREN | WS_TABSTOP | LVS_REPORT | LVS_SHOWSELALWAYS,
                               0,0,0,0,
                               m_hWnd,
                               (HMENU)0,
@@ -323,6 +323,12 @@ public:
 	LRESULT OnCustomDraw(NMHDR *pnmhdr)
 	{
 		NMLVCUSTOMDRAW *pnmlvcd = (NMLVCUSTOMDRAW *)pnmhdr;
+
+		int iGroup = (int)ListView_GetFocusedGroup(m_hWndList);
+		if( iGroup == -1 )
+			SendMessage(m_hWndList,WM_UPDATEUISTATE,MAKELPARAM(UIS_SET,UISF_HIDEFOCUS),0);
+		else
+			SendMessage(m_hWndList,WM_UPDATEUISTATE,MAKELPARAM(UIS_CLEAR,UISF_HIDEFOCUS),0);
 
 		if( pnmlvcd->nmcd.hdr.hwndFrom != m_hWndList )
 			return CDRF_DODEFAULT;